Pulled alert page content to client component
